+++ src/.container.config.php
@@ -27,6 +27,9 @@ use TaskScheduler\WorkerFactoryInterface;
 use Balloon\Filesystem\Node\Factory as NodeFactory;
 use Balloon\Filesystem\Storage\Adapter\Gridfs as GridfsStorage;
 use Balloon\Filesystem\Storage\Adapter\AdapterInterface as StorageAdapterInterface;
+use ParagonIE\Halite\Symmetric\EncryptionKey;
+use ParagonIE\Halite\KeyFactory;
+use ParagonIE\Halite\HiddenString;
 
 return [
     Client::class => [
@@ -45,6 +48,20 @@ return [
         'services' => [
             StorageAdapterInterface::class => [
                 'use' => GridfsStorage::class
+            ]
+        ]
+    ],
+    EncryptionKey::class => [
+        'use' => KeyFactory::class,
+        'factory' => 'importEncryptionKey',
+        'arguments' => [
+            'keyData' => '{'.HiddenString::class.'}'
+        ],
+        'services' => [
+            HiddenString::class => [
+                'arguments' => [
+                    'value' => "{ENV(BALLOON_ENCRYPTION_KEY,3140040033da9bd0dedd8babc8b89cda7f2132dd5009cc43c619382863d0c75e172ebf18e713e1987f35d6ea3ace43b561c50d9aefc4441a8c4418f6928a70e4655de5a9660cd323de63b4fd2fb76525470f25311c788c5e366e29bf60c438c4ac0b440e)}"
+                ]
             ]
         ]
     ],